refactor: separate key guards from property refinements
